Port integration tests to evals.
